+++ Controller/LoginController.php
@@ -15,6 +15,7 @@ use HWI\Bundle\OAuthBundle\Security\Core\Exception\AccountNotLinkedException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
@@ -54,22 +55,22 @@ final class LoginController extends AbstractController
     private $authorizationChecker;
 
     /**
-     * @var SessionInterface
+     * @var RequestStack
      */
-    private $session;
+    private $requestStack;
 
     public function __construct(
         AuthenticationUtils $authenticationUtils,
         RouterInterface $router,
         AuthorizationCheckerInterface $authorizationChecker,
-        SessionInterface $session,
+        RequestStack $requestStack,
         bool $connect,
         string $grantRule
     ) {
         $this->authenticationUtils = $authenticationUtils;
         $this->router = $router;
         $this->authorizationChecker = $authorizationChecker;
-        $this->session = $session;
+        $this->requestStack = $requestStack;
         $this->connect = $connect;
         $this->grantRule = $grantRule;
     }
@@ -93,7 +94,7 @@ final class LoginController extends AbstractController
         // if connecting is enabled and there is no user, redirect to the registration form
         if ($this->connect && !$hasUser && $error instanceof AccountNotLinkedException) {
             $key = time();
-            $session = $request->hasSession() ? $request->getSession() : $this->session;
+            $session = $request->hasSession() ? $request->getSession() : $this->getSession();
             if ($session) {
                 if (!$session->isStarted()) {
                     $session->start();
@@ -112,5 +113,18 @@ final class LoginController extends AbstractController
         return $this->render('@HWIOAuth/Connect/login.html.twig', [
             'error' => $error,
         ]);
+    }
+
+    private function getSession(): ?SessionInterface
+    {
+        if (method_exists($this->requestStack, 'getSession')) {
+            return $this->requestStack->getSession();
+        }
+
+        if ((null !== $request = $this->requestStack->getCurrentRequest()) && $request->hasSession()) {
+            return $request->getSession();
+        }
+
+        return null;
     }
 }