+++ OAuth/RequestDataStorage/SessionStorage.php
@@ -13,6 +13,7 @@ namespace HWI\Bundle\OAuthBundle\OAuth\RequestDataStorage;
 
 use HWI\Bundle\OAuthBundle\OAuth\RequestDataStorageInterface;
 use HWI\Bundle\OAuthBundle\OAuth\ResourceOwnerInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 /**
@@ -21,17 +22,19 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
  * @author Alexander <iam.asm89@gmail.com>
  * @author Francisco Facioni <fran6co@gmail.com>
  * @author Joseph Bielawski <stloyd@gmail.com>
+ *
+ * @final since 1.4
  */
 class SessionStorage implements RequestDataStorageInterface
 {
     /**
-     * @var SessionInterface
+     * @var RequestStack
      */
-    private $session;
+    private $requestStack;
 
-    public function __construct(SessionInterface $session)
+    public function __construct(RequestStack $requestStack)
     {
-        $this->session = $session;
+        $this->requestStack = $requestStack;
     }
 
     /**
@@ -40,13 +43,13 @@ class SessionStorage implements RequestDataStorageInterface
     public function fetch(ResourceOwnerInterface $resourceOwner, $key, $type = 'token')
     {
         $key = $this->generateKey($resourceOwner, $key, $type);
-        if (null === $data = $this->session->get($key)) {
+        if (null === $data = $this->getSession()->get($key)) {
             throw new \InvalidArgumentException('No data available in storage.');
         }
 
-        // request tokens are one time use only
-        if (\in_array($type, ['token', 'csrf_state'])) {
-            $this->session->remove($key);
+        // Request tokens are one time use only
+        if (\in_array($type, ['token', 'csrf_state'], true)) {
+            $this->getSession()->remove($key);
         }
 
         return $data;
@@ -67,7 +70,7 @@ class SessionStorage implements RequestDataStorageInterface
             $key = $this->generateKey($resourceOwner, $this->getStorageKey($value), $type);
         }
 
-        $this->session->set($key, $this->getStorageValue($value));
+        $this->getSession()->set($key, $this->getStorageValue($value));
     }
 
     /**
@@ -111,5 +114,18 @@ class SessionStorage implements RequestDataStorageInterface
         }
 
         return (string) $storageKey;
+    }
+
+    private function getSession(): SessionInterface
+    {
+        if (method_exists($this->requestStack, 'getSession')) {
+            return $this->requestStack->getSession();
+        }
+
+        if ((null !== $request = $this->requestStack->getCurrentRequest()) && $request->hasSession()) {
+            return $request->getSession();
+        }
+
+        throw new \LogicException('There is currently no session available.');
     }
 }