+++ Tests/OAuth/RequestDataStorage/SessionStorageTest.php
@@ -15,6 +15,8 @@ use HWI\Bundle\OAuthBundle\OAuth\RequestDataStorage\SessionStorage;
 use HWI\Bundle\OAuthBundle\OAuth\ResourceOwnerInterface;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 class SessionStorageTest extends TestCase
@@ -41,7 +43,18 @@ class SessionStorageTest extends TestCase
         $this->resourceOwner->method('getName')->willReturn('resource_owner_name');
         $this->resourceOwner->method('getOption')->with('client_id')->willReturn('client_id');
 
-        $this->storage = new SessionStorage($this->session);
+        $request = $this->createMock(Request::class);
+        $request->method('hasSession')->willReturn(true);
+        $request->method('getSession')->willReturn($this->session);
+
+        $requestStack = $this->createMock(RequestStack::class);
+        $requestStack->method('getCurrentRequest')->willReturn($request);
+
+        if (method_exists(RequestStack::class, 'getSession')) {
+            $requestStack->method('getSession')->willReturn($this->session);
+        }
+
+        $this->storage = new SessionStorage($requestStack);
     }
 
     public function testSaveTokenWithoutOAuthTokenPassedThrowsException(): void