+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 Changelog
 =========
 ## 1.4.0 (2021-xx-xx)
-* Dropped Symfony 5.0 support as it is EOL,
-* Dropped PHP 7.2 support as it is EOL,
+* BC Break: dropped Symfony 5.0 support as it is EOL,
+* BC Break: dropped PHP 7.2 support as it is EOL,
+* BC Break: changed `__construct()` argument for `OAuth/RequestDataStorage/SessionStorage`, from `SessionInterface $session` to `RequestStack $requestStack`,
 * Added: Symfony 5.1 Security system support,
 * Added: state support for service authentication URL's,
 * Added: ability to change the response after `HWIOAuthEvents::CONNECT_COMPLETED` is fired,